fix lock not released if lock callback throws
There was a bug where the user-supplied callback to the request()
function throws an exception, the lock is not released. This was due
to the fact that the promise in _resolveWithCB() never resolves or
rejects if the wrapped promise rejects.

This fixes the issue by resolving even in the case that the user
provided promise rejects.

A test is added the was failing before this fix was made and succeeds
after the fix is made.